+++ payments/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "jest --watchAll --no-cache",
-    "start": "ts-node-dev --poll src/payments.ts"
+    "start": "ts-node-dev --poll src/payments.ts",
+    "test:ci": "jest"
   },
   "jest": {
     "preset": "ts-jest",
