+++ tickets/src/events/listeners/__test__/orderCreaLstnr.test.ts
@@ -1,18 +1,17 @@
 import {natsWrapper} from  '../../../natsWrapper';
-import {OrderCancListener} from  '../../listeners/orderCancLstnr';
+import {OrderCreatedListener} from  '../orderCreaLstnr';
 import {Ticket} from '../../../dbmodel/ticket';
-import {OrderCancelledEvent, OrderStatus} from '@tk-test-org/tk-test-common';
+import {OrderCreatedEvent, OrderStatus} from '@tk-test-org/tk-test-common';
 import mongoose from 'mongoose';
 import { Message } from 'node-nats-streaming';
 
 const setup = async () => {
-    const myListener = new OrderCancListener(natsWrapper.client);
+    const myListener = new OrderCreatedListener(natsWrapper.client);
     const tick = Ticket.build({title: 'fist title', price: 10, userId: '12345'});
-    tick.set({ orderId: mongoose.Types.ObjectId().toHexString() });
     await tick.save();
 
     //create fake data
-    const data:OrderCancelledEvent['data'] = {
+    const data:OrderCreatedEvent['data'] = {
         id: mongoose.Types.ObjectId().toHexString(),
         version: 0,
         userId: '12345',
@@ -32,12 +31,23 @@ const setup = async () => {
     return {myListener, tick, data, msg};
 };
 
-it(' updates ticket, publishes update and acks nats message  ', async () => {
+it(' sets order id for ticket ', async () => {
     const {myListener, tick, data, msg} = await setup();
     await myListener.onMessage(data, msg);
     const myTick = await Ticket.findById(tick.id);
-    //expect(myTick.orderId).toEqual(undefined);
-    expect(myTick.orderId).not.toBeDefined();
+    expect(myTick.orderId).toEqual(data.id);
+});
+
+it(' acks the message ', async () => {
+    const {myListener, tick, data, msg} = await setup();
+    //console.log('tick.id', tick.id);
+    await myListener.onMessage(data, msg);
     expect(msg.ack).toHaveBeenCalled();
+});
+
+it(' publishes ticket updated event ', async ()=> {
+    const {myListener, tick, data, msg} = await setup();
+    await myListener.onMessage(data, msg);
     expect(natsWrapper.client.publish).toHaveBeenCalled();
+    //console.log((natsWrapper.client.publish as jest.Mock).mock.calls);
 });
